fix: avoid server fetchers in client components
